requirements.txt
aiogram>=3.4.1
google-generativeai>=0.4.1
python-dotenv>=1.0.1
aiosqlite>=0.20.0
cachetools>=5.5.0
